requirements.txt
streamlit
pandas
langchain
groq
faiss-cpu
python-dotenv
langchain-community
langchain-huggingface
sentence-transformers